+++ tests/inputs/return-hanging-expression-2.lua
@@ -1,4 +1,6 @@
-return cframe * Vector3.new( -- Clamp & transform into world space
+return cframe
+	-- Clamp & transform into world space
+	* Vector3.new(
 		math.clamp(transform.X, -halfSize.X, halfSize.X),
 		math.clamp(transform.Y, -halfSize.Y, halfSize.Y),
 		math.clamp(transform.Z, -halfSize.Z, halfSize.Z)