+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed an anonymous function assignment `local x = function()` being unnecessarily indented if the function body contains a comment
 - Fixed malformed formatting when there is a newline between a `return` token and the expressions
+- Fixed malformed formatting of multi-assignment or multi-returns where there is a comment within the expressions list ([#637](https://github.com/JohnnyMorganz/StyLua/issues/637))
 
 ## [0.15.3] - 2022-12-07